feat(manga): filter extension pickers by selected languages
